% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agePredHorvath353}
\alias{agePredHorvath353}
\title{353 age predictor probes from Horvath 2013}
\format{An object of class \code{data.frame} with 354 rows and 23 columns.}
\usage{
agePredHorvath353
}
\description{
353 age predictor probes from Horvath 2013
}
\keyword{datasets}
