% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.CD14Monocytes}
\alias{cellref.CD14Monocytes}
\title{cellref.CD14Monocytes}
\format{An object of class \code{numeric} of length 351705.}
\usage{
cellref.CD14Monocytes
}
\description{
cellref.CD14Monocytes
}
\keyword{datasets}
