% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.CD19B}
\alias{cellref.CD19B}
\title{cellref.CD19B}
\format{An object of class \code{numeric} of length 342174.}
\usage{
cellref.CD19B
}
\description{
Vector of 342174 binarized DNA methylation status for CD19B cells.
}
\details{
Beta values were first trichotomized by <=0.3 or >=0.7. A methylation
status is determined if 80% of the samples show consistent calls and
no conflicting calls exist.

Original data source: GSE35069
}
\keyword{datasets}
