% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.CD4T}
\alias{cellref.CD4T}
\title{cellref.CD4T}
\format{An object of class \code{numeric} of length 343658.}
\usage{
cellref.CD4T
}
\description{
Vector of 343658 binarized DNA methylation status for CD4T cells.
}
\keyword{datasets}
