% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.CD56NK}
\alias{cellref.CD56NK}
\title{cellref.CD56NK}
\format{An object of class \code{numeric} of length 336647.}
\usage{
cellref.CD56NK
}
\description{
cellref.CD56NK
}
\keyword{datasets}
