% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.CD8T}
\alias{cellref.CD8T}
\title{cellref.CD8T}
\format{An object of class \code{numeric} of length 332973.}
\usage{
cellref.CD8T
}
\description{
Vector of 332973 binarized DNA methylation status for CD8T cells.
}
\details{
Beta values were first trichotomized by <=0.3 or >=0.7. A methylation
status is determined if 80% of the samples show consistent calls and
no conflicting calls exist.

Original data source: GSE35069
}
\keyword{datasets}
