% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.eosinophil}
\alias{cellref.eosinophil}
\title{cellref.eosinophil}
\format{An object of class \code{numeric} of length 342377.}
\usage{
cellref.eosinophil
}
\description{
Vector of 342377 binarized DNA methylation status for eosinophil cells.
}
\keyword{datasets}
