% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.hair}
\alias{cellref.hair}
\title{cellref.hair}
\format{An object of class \code{numeric} of length 317041.}
\usage{
cellref.hair
}
\description{
Vector of 317041 binarized DNA methylation status for hair cells.
}
\keyword{datasets}
