% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.liver}
\alias{cellref.liver}
\title{cellref.liver}
\format{An object of class \code{numeric} of length 292658.}
\usage{
cellref.liver
}
\description{
Vector of 292658 binarized DNA methylation status for liver cells.
}
\keyword{datasets}
