% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.muscle}
\alias{cellref.muscle}
\title{cellref.muscle}
\format{An object of class \code{numeric} of length 292577.}
\usage{
cellref.muscle
}
\description{
Vector of 292577 binarized DNA methylation status for muscle cells.
}
\details{
Beta values were first trichotomized by <=0.3 or >=0.7. A methylation
status is determined if 80% of the samples show consistent calls and
no conflicting calls exist.

Original data source: GSE48472
}
\keyword{datasets}
