% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.scFat}
\alias{cellref.scFat}
\title{cellref.scFat}
\format{An object of class \code{numeric} of length 283869.}
\usage{
cellref.scFat
}
\description{
Vector of 283869 binarized DNA methylation status for scFat cells.
}
\details{
Beta values were first trichotomized by <=0.3 or >=0.7. A methylation
status is determined if 80% of the samples show consistent calls and
no conflicting calls exist.

Original data source: GSE48472
}
\keyword{datasets}
