% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.spleen}
\alias{cellref.spleen}
\title{cellref.spleen}
\format{An object of class \code{numeric} of length 302593.}
\usage{
cellref.spleen
}
\description{
Vector of 302593 binarized DNA methylation status for spleen cells.
}
\keyword{datasets}
