% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoBand.hg19}
\alias{cytoBand.hg19}
\title{Cytoband information of hg19}
\format{An object of class \code{data.frame} with 862 rows and 5 columns.}
\usage{
cytoBand.hg19
}
\description{
Cytoband coordinates for genome build hg19. Data frame with columns
chrom, chromStart, chromEnd, name and geiStain.
}
\keyword{datasets}
