% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoBand.hg38}
\alias{cytoBand.hg38}
\title{Cytoband information of hg38}
\format{An object of class \code{data.frame} with 1293 rows and 5 columns.}
\usage{
cytoBand.hg38
}
\description{
Cytoband coordinates for genome build hg38. Data frame with columns
chrom, chromStart, chromEnd, name and geiStain.
}
\keyword{datasets}
