% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.female.clean.chrY.probes}
\alias{EPIC.female.clean.chrY.probes}
\title{EPIC.female.clean.chrY.probes}
\format{An object of class \code{character} of length 314.}
\usage{
EPIC.female.clean.chrY.probes
}
\description{
Vector of 314 probe IDs for chromosome Y probes in EPIC array excluding
pseudoautosomal regions. Pseudoautosomal regions was identified based on
cross-hybridization signal in TCGA normal samples.
}
\details{
See scripts/make-data.R for processing details
}
\keyword{datasets}
