% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.female.xlinked.chrX.probes}
\alias{EPIC.female.xlinked.chrX.probes}
\title{EPIC.female.xlinked.chrX.probes}
\format{An object of class \code{character} of length 3433.}
\usage{
EPIC.female.xlinked.chrX.probes
}
\description{
Vector of 3433 probe IDs for X-linked probes in EPIC array
}
\keyword{datasets}
