% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.ordering}
\alias{EPIC.ordering}
\title{EPIC probe ordering}
\format{An object of class \code{data.frame} with 865918 rows and 6 columns.}
\usage{
EPIC.ordering
}
\description{
EPIC probe chip address
}
\keyword{datasets}
