% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.typeI.extC}
\alias{EPIC.typeI.extC}
\title{EPIC.typeI.extC}
\format{An object of class \code{character} of length 46733.}
\usage{
EPIC.typeI.extC
}
\description{
Vector of 46733 C-extension Type-I probes included in EPIC array.
}
\details{
Source: http://zwdzwd.github.io/InfiniumAnnotation
}
\keyword{datasets}
