% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.typeI.extT}
\alias{EPIC.typeI.extT}
\title{EPIC.typeI.extT}
\format{An object of class \code{character} of length 15638.}
\usage{
EPIC.typeI.extT
}
\description{
Vector of 15638 T-extension Type-I probes included in EPIC array.
}
\details{
Source: http://zwdzwd.github.io/InfiniumAnnotation
}
\keyword{datasets}
