% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethnicity.ccs.probes}
\alias{ethnicity.ccs.probes}
\title{ethnicity.ccs.probes}
\format{An object of class \code{character} of length 332.}
\usage{
ethnicity.ccs.probes
}
\description{
ethnicity.ccs.probes
}
\keyword{datasets}
