% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethnicity.rs.probes}
\alias{ethnicity.rs.probes}
\title{ethnicity.rs.probes}
\format{An object of class \code{character} of length 59.}
\usage{
ethnicity.rs.probes
}
\description{
Vector of 59 built-in SNP probes for ethnicity inference.
}
\details{
Source: http://zwdzwd.github.io/InfiniumAnnotation
}
\keyword{datasets}
