% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19.chrominfo}
\alias{hg19.chrominfo}
\title{hg19.chrominfo}
\format{An object of class \code{list} of length 2.}
\usage{
hg19.chrominfo
}
\description{
List of two GenomicRange objects. The first represents the genomic
coordinates for gap, the second for sequence lengths. The information
is intended for genome build hg19.
}
\keyword{datasets}
