% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.chrominfo}
\alias{hg38.chrominfo}
\title{hg38.chrominfo}
\format{An object of class \code{list} of length 2.}
\usage{
hg38.chrominfo
}
\description{
List of two GenomicRange objects. The first represents the genomic
coordinates for gap, the second for sequence lengths. The information
is intended for genome build hg38.
}
\details{
See scripts/make-data.R for building details.
}
\keyword{datasets}
