% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM27.betas.leuko.whole}
\alias{HM27.betas.leuko.whole}
\title{HM27.betas.leuko.whole}
\format{An object of class \code{data.frame} with 27578 rows and 2 columns.}
\usage{
HM27.betas.leuko.whole
}
\description{
Matrix of 27578 probes x 2 samples of whole blood DNA methylation
beta values. Data obtained from TCGA pilot study.
}
\details{
Preprocessing involves background subtraction and dye bias correction.
See scripts/make-data.R for building details.
}
\keyword{datasets}
