% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM27.controls}
\alias{HM27.controls}
\title{HM27.controls}
\format{An object of class \code{data.frame} with 144 rows and 4 columns.}
\usage{
HM27.controls
}
\description{
HM27.controls
}
\keyword{datasets}
