% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM27.hg19.probe2chr}
\alias{HM27.hg19.probe2chr}
\title{HM27.hg19.probe2chr}
\format{An object of class \code{factor} of length 27588.}
\usage{
HM27.hg19.probe2chr
}
\description{
Named vector that maps 27588 HM27 probe to chromosomes based on genome
build hg19.
}
\keyword{datasets}
