% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM27.ordering}
\alias{HM27.ordering}
\title{HM27 probe ordering}
\format{An object of class \code{data.frame} with 27578 rows and 6 columns.}
\usage{
HM27.ordering
}
\description{
Data frame of 27578 rows each describing the chip address of a probe.
Column names include Probe ID, methylated allele address, unmethylated
allele address, probe design type, color channel and abbreviations.
}
\details{
Source: http://zwdzwd.github.io/InfiniumAnnotation

HM27 probe chip address
}
\keyword{datasets}
