% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.betas.leuko.whole}
\alias{HM450.betas.leuko.whole}
\title{HM450.betas.leuko.whole}
\format{An object of class \code{data.frame} with 485577 rows and 3 columns.}
\usage{
HM450.betas.leuko.whole
}
\description{
HM450.betas.leuko.whole
}
\keyword{datasets}
