% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.female.xlinked.chrX.probes}
\alias{HM450.female.xlinked.chrX.probes}
\title{HM450.female.xlinked.chrX.probes}
\format{An object of class \code{character} of length 3797.}
\usage{
HM450.female.xlinked.chrX.probes
}
\description{
Vector of 3797 probe IDs for X-linked probes in HM450 array.  X-linkage is
based on intermediate DNA methylation signal in TCGA normal female samples.
}
\details{
See scripts/make-data.R for processing details
}
\keyword{datasets}
