% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.hg19.probe2chr}
\alias{HM450.hg19.probe2chr}
\title{HM450.hg19.probe2chr}
\format{An object of class \code{factor} of length 485577.}
\usage{
HM450.hg19.probe2chr
}
\description{
Named vector that maps 485577 HM450 probe to chromosomes based on genome
build hg19.
}
\keyword{datasets}
