% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.mapped.probes.hg19}
\alias{HM450.mapped.probes.hg19}
\title{HM450.mapped.probes.hg19}
\format{An object of class \code{GRanges} of length 467097.}
\usage{
HM450.mapped.probes.hg19
}
\description{
GenomicRange object for 467097 CpG probes included in EPIC array based
on genome build hg19.
}
\details{
Source: http://zwdzwd.github.io/InfiniumAnnotation
}
\keyword{datasets}
