% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.mapped.probes.hg38}
\alias{HM450.mapped.probes.hg38}
\title{HM450.mapped.probes.hg38}
\format{An object of class \code{GRanges} of length 466007.}
\usage{
HM450.mapped.probes.hg38
}
\description{
HM450.mapped.probes.hg38
}
\keyword{datasets}
