% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.mask.tcga}
\alias{HM450.mask.tcga}
\title{HM450.mask.tcga}
\format{An object of class \code{character} of length 89512.}
\usage{
HM450.mask.tcga
}
\description{
HM450.mask.tcga
}
\keyword{datasets}
