% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.ordering}
\alias{HM450.ordering}
\title{HM450 probe ordering}
\format{An object of class \code{data.frame} with 485577 rows and 6 columns.}
\usage{
HM450.ordering
}
\description{
HM450 probe chip address
}
\keyword{datasets}
