% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.typeI.extC}
\alias{HM450.typeI.extC}
\title{HM450.typeI.extC}
\format{An object of class \code{character} of length 45427.}
\usage{
HM450.typeI.extC
}
\description{
HM450.typeI.extC
}
\keyword{datasets}
