% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.typeI.extT}
\alias{HM450.typeI.extT}
\title{HM450.typeI.extT}
\format{An object of class \code{character} of length 15148.}
\usage{
HM450.typeI.extT
}
\description{
HM450.typeI.extT
}
\keyword{datasets}
