% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno.R
\name{sesameData_getAnno}
\alias{sesameData_getAnno}
\title{Retrieve additional annotation files}
\usage{
sesameData_getAnno(title, version = anno_base_default_version, dest_dir = NULL)
}
\arguments{
\item{title}{title of the annotation file}

\item{version}{version number}

\item{dest_dir}{if not NULL, download to this directory}
}
\value{
annotation file
}
\description{
From the Infinium annotation website associated github repo
e.g., 
https://github.com/zhou-lab/InfiniumAnnotationV1
}
\details{
The default version number should always work. One need to
refer to the actual repo to see which one of the other versions
also work.

See also
http://zwdzwd.github.io/InfiniumAnnotation
}
\examples{

## avoided testing these function as they use external resources
if (FALSE) {
mft <- sesameData_getAnno("HM27/HM27.hg19.manifest.tsv.gz")
annoI <- sesameData_getAnno("EPIC/EPIC.hg19.typeI_overlap_b151.rds")
annoS <- sesameData_getAnno("EPIC/EPIC.hg19.snp_overlap_b151.rds")
sesameData_getAnno("test/3999492009_R01C01_Grn.idat", dest_dir = tempdir())
}

}
