% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbes.R
\name{sesameData_getAutosomeProbes}
\alias{sesameData_getAutosomeProbes}
\title{Get autosome probes}
\usage{
sesameData_getAutosomeProbes(platform = NULL, genome = NULL)
}
\arguments{
\item{platform}{'EPIC', 'HM450' etc.}

\item{genome}{hg19, hg38, or mm10, inference by default}
}
\value{
GRanges of autosome probes
}
\description{
Get autosome probes
}
\examples{
auto_probes <- sesameData_getAutosomeProbes('Mammal40')
}
