% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\name{sesameData_getGeneGRanges}
\alias{sesameData_getGeneGRanges}
\title{convert transcript GRanges to gene GRanges}
\usage{
sesameData_getGeneGRanges(txns)
}
\arguments{
\item{txns}{GRanges object}
}
\value{
a GRanges object
}
\description{
convert transcript GRanges to gene GRanges
}
\examples{
full_model <- sesameDataGet("genomeInfo.mm10")$txns
txns <- sesameData_getTxnGRanges(full_model)
genes <- sesameData_getGeneGRanges(txns)

}
