% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{sesameData_getManifestDF}
\alias{sesameData_getManifestDF}
\title{download Infinium manifest from Github repositories}
\usage{
sesameData_getManifestDF(platform, genome = NULL, version = 1)
}
\arguments{
\item{platform}{Mammal40, MM285, EPIC, and HM450}

\item{genome}{hg38, mm10 etc.}

\item{version}{release version, default is the latest}
}
\value{
tibble
}
\description{
download Infinium manifest from Github repositories
}
\examples{
mft <- sesameData_getManifestDF("Mammal40")
}
