% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{sesameData_getManifestGRanges}
\alias{sesameData_getManifestGRanges}
\title{get Infinium manifest GRanges}
\usage{
sesameData_getManifestGRanges(platform, genome = NULL)
}
\arguments{
\item{platform}{Mammal40, MM285, EPIC, and HM450}

\item{genome}{hg38, mm10 etc.}
}
\value{
GRanges
}
\description{
for building manifest from the original
}
\examples{
gr <- sesameData_getManifestGRanges("Mammal40")
}
