% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbes.R
\name{sesameData_getProbesByChromosome}
\alias{sesameData_getProbesByChromosome}
\title{Get Probes by Chromosome}
\usage{
sesameData_getProbesByChromosome(chrms, platform = NULL, genome = NULL)
}
\arguments{
\item{chrms}{chromosomes to subset}

\item{platform}{EPIC, HM450, Mouse}

\item{genome}{hg38, mm10, ... will infer if not given.
For additional mapping, download the GRanges object from
http://zwdzwd.github.io/InfiniumAnnotation
and provide the following argument
..., genome = sesameAnno_buildManifestGRanges("downloaded_file"),...
to this function.}
}
\value{
GRanges of selected probes
}
\description{
Get Probes by Chromosome
}
\examples{
Xprobes <- sesameData_getProbesByChromosome('chrX', "Mammal40")
}
