% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbes.R
\name{sesameData_getProbesByRegion}
\alias{sesameData_getProbesByRegion}
\title{Get probes by genomic region}
\usage{
sesameData_getProbesByRegion(
  regs,
  chrm = NULL,
  beg = 1,
  end = -1,
  platform = NULL,
  genome = NULL
)
}
\arguments{
\item{regs}{GRanges}

\item{chrm}{chromosome, when given regs are ignored}

\item{beg}{begin, 1 if omitted}

\item{end}{end, chromosome end if omitted}

\item{platform}{EPIC, HM450, ...}

\item{genome}{hg38, mm10, ... will infer if not given.
For additional mapping, download the GRanges object from
http://zwdzwd.github.io/InfiniumAnnotation
and provide the following argument
..., genome = sesameAnno_buildManifestGRanges("downloaded_file"),...
to this function.}
}
\value{
GRanges of selected probes
}
\description{
The function takes a genomic coordinate and output the a vector of probes
on the specified platform that falls in the given genomic region.
}
\examples{
library(GenomicRanges)
sesameData_getProbesByRegion(
    GRanges('chr5', IRanges(135313937, 135419936)), platform = 'Mammal40')
}
