% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbes.R
\name{sesameData_getProbesByTSS}
\alias{sesameData_getProbesByTSS}
\title{Get Probes by Gene Transcription Start Site (TSS)}
\usage{
sesameData_getProbesByTSS(
  gene_name,
  platform = NULL,
  upstream = 1500,
  downstream = 1500,
  genome = NULL
)
}
\arguments{
\item{gene_name}{gene name}

\item{platform}{EPIC, HM450, or MM285}

\item{upstream}{the number of base pairs to expand upstream the TSS}

\item{downstream}{the number of base pairs to expand downstream the TSS}

\item{genome}{hg38, hg19 or mm10}
}
\value{
probes that fall into the given gene
}
\description{
Get probes mapped to a TSS. All transcripts for the gene are considered.
The function takes a gene name as appears in UCSC RefGene database. The
platform and reference genome build can be changed with `platform` and
`genome` options. The function returns a vector of probes that falls
into the TSS region of the gene.
}
\examples{
probes <- sesameData_getProbesByTSS('DNMT3A', "Mammal40")
}
