% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\name{sesameData_getTranscriptsByGene}
\alias{sesameData_getTranscriptsByGene}
\title{get transcripts by gene name}
\usage{
sesameData_getTranscriptsByGene(gene_name, genome)
}
\arguments{
\item{gene_name}{gene name}

\item{genome}{hg38, mm10 etc.}
}
\value{
GRangesList of the transcripts
}
\description{
get transcripts by gene name
}
\examples{
sesameData_getTranscriptsByGene("Cdkn2a", "mm10")
}
