% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{sesameData_getTxnGRanges}
\alias{sesameData_getTxnGRanges}
\title{convert GRangesList to transcript GRanges}
\usage{
sesameData_getTxnGRanges(genome = NULL, grl = NULL, merge2gene = FALSE)
}
\arguments{
\item{genome}{hg38, mm10, ...}

\item{grl}{GRangesList object}

\item{merge2gene}{merge transcript to genes}
}
\value{
a GRanges object
}
\description{
convert GRangesList to transcript GRanges
}
\examples{
txns <- sesameData_getTxnGRanges("mm10")
## get verified protein-coding
txns <- txns[(txns$transcript_type == "protein_coding" & txns$level <= 2)]
genes <- sesameData_getTxnGRanges(txns, merge2gene = TRUE)
}
