% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sesameDataCache}
\alias{sesameDataCache}
\title{Cache SeSAMe data for specific platform}
\usage{
sesameDataCache(platform = NULL, keyword = NULL)
}
\arguments{
\item{platform}{EPIC, HM450, MM285, etc.}

\item{keyword}{keyword used to filter records}
}
\value{
TRUE
}
\description{
Cache SeSAMe data for specific platform
}
\examples{
if(FALSE) { sesameDataCache("MM285") }
}
