% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sesameDataCacheExample}
\alias{sesameDataCacheExample}
\title{Cache SeSAMe data for specific platform}
\usage{
sesameDataCacheExample()
}
\value{
TRUE
}
\description{
Cache SeSAMe data for specific platform
}
\examples{
if(FALSE) { sesameDataCacheExample() }
}
