% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sesameDataClearCache}
\alias{sesameDataClearCache}
\title{Clear cache to free memory}
\usage{
sesameDataClearCache()
}
\value{
gc() output
}
\description{
sesameData:::cacheEnv
}
\examples{
sesameDataClearCache()
}
