% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameDataDownload.R
\name{sesameDataDownload}
\alias{sesameDataDownload}
\title{Download auxiliary data for sesame function and documentation}
\usage{
sesameDataDownload(file_name, dest_dir = NULL)
}
\arguments{
\item{file_name}{name of file to download}

\item{dest_dir}{directory to hold downloaded file.
use the temporary directory if not given}
}
\value{
a list with url, dest_dir, dest_file and file_name
}
\description{
Download auxiliary data for sesame function and documentation
}
\examples{

## We avoided testing this function to ensure version consistency.
if(FALSE) { sesameDataDownload("3999492009_R01C01_Grn.idat") }
}
