% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{sesameDataDownloadAnno}
\alias{sesameDataDownloadAnno}
\title{Retrieve manifest file from the supporting website
at http://zwdzwd.github.io/InfiniumAnnotation
and https://github.com/zhou-lab/InfiniumAnnotation}
\usage{
sesameDataDownloadAnno(title)
}
\arguments{
\item{title}{title of the annotation file}
}
\value{
annotation file
}
\description{
Retrieve manifest file from the supporting website
at http://zwdzwd.github.io/InfiniumAnnotation
and https://github.com/zhou-lab/InfiniumAnnotation
}
\examples{

mft <- sesameDataDownloadAnno("HM450/HM450.hg19.manifest")

}
