% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sesameDataGet}
\alias{sesameDataGet}
\title{Get SeSAMe data}
\usage{
sesameDataGet(title)
}
\arguments{
\item{title}{title of the data}
}
\value{
data object
}
\description{
Get SeSAMe data
}
\examples{

sesameDataGet('genomeInfo.hg38')
}
