% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{sesameDataGetAnno}
\alias{sesameDataGetAnno}
\title{Retrieve manifest file from the supporting website
at http://zwdzwd.github.io/InfiniumAnnotation
and https://github.com/zhou-lab/InfiniumAnnotation}
\usage{
sesameDataGetAnno(title)
}
\arguments{
\item{title}{title of the annotation file}
}
\value{
annotation file
}
\description{
Retrieve manifest file from the supporting website
at http://zwdzwd.github.io/InfiniumAnnotation
and https://github.com/zhou-lab/InfiniumAnnotation
}
\examples{

mft <- sesameDataGetAnno("HM27/HM27.hg19.manifest.tsv.gz")
annoS <- sesameDataGetAnno("EPIC/EPIC.hg19.typeI_overlap_b151.rds")

}
