% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sesameDataList}
\alias{sesameDataList}
\title{List all SeSAMe data}
\usage{
sesameDataList(dateAdded = latest_date)
}
\arguments{
\item{dateAdded}{version of the data by date added, if "all", show all dates}
}
\value{
all titles from SeSAMe Data
}
\description{
List all SeSAMe data
}
\examples{
sesameDataList()
}
