% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sesameDataListDates}
\alias{sesameDataListDates}
\title{List all versions of SeSAMe data}
\usage{
sesameDataListDates()
}
\value{
sorted unique dates of SeSAMe Data
}
\description{
List all versions of SeSAMe data
}
\examples{
sesameDataListDates()
}
