% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex.inference.model}
\alias{sex.inference.model}
\title{sex.inference.model}
\format{An object of class \code{randomForest} of length 18.}
\usage{
sex.inference.model
}
\description{
Random forest model trained for sex inference. An object of class
"randomForest". The training data set is 719 TCGA adjacent normals.
}
\details{
See scripts/make-data.R for details
}
\keyword{datasets}
