% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC.refGene.gene2txn.hg19}
\alias{UCSC.refGene.gene2txn.hg19}
\title{UCSC.refGene.gene2txn.hg19}
\format{An object of class \code{list} of length 27129.}
\usage{
UCSC.refGene.gene2txn.hg19
}
\description{
UCSC.refGene.gene2txn.hg19
}
\keyword{datasets}
