% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC.refGene.gene2txn.hg38}
\alias{UCSC.refGene.gene2txn.hg38}
\title{UCSC.refGene.gene2txn.hg38}
\format{An object of class \code{list} of length 27221.}
\usage{
UCSC.refGene.gene2txn.hg38
}
\description{
UCSC.refGene.gene2txn.hg38
}
\keyword{datasets}
