% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC.refGene.txn2gene.hg19}
\alias{UCSC.refGene.txn2gene.hg19}
\title{UCSC.refGene.txn2gene.hg19}
\format{An object of class \code{list} of length 57956.}
\usage{
UCSC.refGene.txn2gene.hg19
}
\description{
UCSC.refGene.txn2gene.hg19
}
\keyword{datasets}
