% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC.refGene.txns.hg19}
\alias{UCSC.refGene.txns.hg19}
\title{UCSC.refGene.txns.hg19}
\format{An object of class \code{GRangesList} of length 62083.}
\usage{
UCSC.refGene.txns.hg19
}
\description{
GRangesList object of length 62083. Each contains a transcript and
the coordinates of its exons. The data was adapted from UCSC refGene
(2016) for genome build hg19.
}
\details{
See scripts/make-data.R for building details.
}
\keyword{datasets}
