% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agePredHorvath353}
\alias{agePredHorvath353}
\title{Age Predictor based on Horvath 2013}
\format{An object of class \code{data.frame} with 354 rows and 23 columns.}
\usage{
agePredHorvath353
}
\description{
Data frame of 353 rows each corresponding a probe. Columns include
probe IDs and regression coefficients.
The data is adapted from Horvath 2013 Genome Biology.
}
\keyword{datasets}
