% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.buccal}
\alias{cellref.buccal}
\title{cellref.buccal}
\format{An object of class \code{numeric} of length 299850.}
\usage{
cellref.buccal
}
\description{
Vector of 299850 binarized DNA methylation status for buccal cells.
}
\keyword{datasets}
