% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.CD14Monocytes}
\alias{cellref.CD14Monocytes}
\title{cellref.CD14Monocytes}
\format{An object of class \code{numeric} of length 351705.}
\usage{
cellref.CD14Monocytes
}
\description{
Vector of 351705 binarized DNA methylation status for CD14Monocytes cells.
Beta values were first trichotomized by <=0.3 or >=0.7. A methylation
status is determined if 80% of the samples show consistent calls and
no conflicting calls exist.
}
\details{
Original data source: GSE35069
}
\keyword{datasets}
