% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.CD56NK}
\alias{cellref.CD56NK}
\title{cellref.CD56NK}
\format{An object of class \code{numeric} of length 336647.}
\usage{
cellref.CD56NK
}
\description{
Vector of 336647 binarized DNA methylation status for CD56NK cells.
}
\details{
Beta values were first trichotomized by <=0.3 or >=0.7. A methylation
status is determined if 80% of the samples show consistent calls and
no conflicting calls exist.

Original data source: GSE35069
}
\keyword{datasets}
