% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.granulocytes}
\alias{cellref.granulocytes}
\title{cellref.granulocytes}
\format{An object of class \code{numeric} of length 347903.}
\usage{
cellref.granulocytes
}
\description{
Vector of 347903 binarized DNA methylation status for granulocytes cells.
}
\keyword{datasets}
