% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.neutrophil}
\alias{cellref.neutrophil}
\title{cellref.neutrophil}
\format{An object of class \code{numeric} of length 343822.}
\usage{
cellref.neutrophil
}
\description{
cellref.neutrophil
}
\keyword{datasets}
