% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.saliva}
\alias{cellref.saliva}
\title{cellref.saliva}
\format{An object of class \code{numeric} of length 323797.}
\usage{
cellref.saliva
}
\description{
Vector of 323797 binarized DNA methylation status for saliva cells.
}
\details{
Beta values were first trichotomized by <=0.3 or >=0.7. A methylation
status is determined if 80% of the samples show consistent calls and
no conflicting calls exist.

Original data source: GSE48472
}
\keyword{datasets}
