% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellref.skin}
\alias{cellref.skin}
\title{cellref.skin}
\format{An object of class \code{numeric} of length 280657.}
\usage{
cellref.skin
}
\description{
Vector of 280657 binarized DNA methylation status for skin cells.
}
\keyword{datasets}
