% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.betas.leuko.whole}
\alias{EPIC.betas.leuko.whole}
\title{EPIC.betas.leuko.whole}
\format{An object of class \code{data.frame} with 485577 rows and 3 columns.}
\usage{
EPIC.betas.leuko.whole
}
\description{
EPIC.betas.leuko.whole
}
\keyword{datasets}
