% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.mask}
\alias{EPIC.mask}
\title{EPIC.mask}
\format{An object of class \code{character} of length 82108.}
\usage{
EPIC.mask
}
\description{
Vector of 82108 probe IDs recommended for masking in EPIC array based on
Zhou et al. 2017 Nucleic Acids Research. MASK.general column of the
original annotation is used here.
}
\details{
Source: http://zwdzwd.github.io/InfiniumAnnotation
}
\keyword{datasets}
