% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethnicity.model}
\alias{ethnicity.model}
\title{ethnicity.model}
\format{An object of class \code{randomForest} of length 18.}
\usage{
ethnicity.model
}
\description{
Random forest model trained for ethnicity inference. An object of class
"randomForest". Details in Zhou et al. 2017 Nucleic Acids Research
}
\keyword{datasets}
