% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM27.betas.leuko.whole}
\alias{HM27.betas.leuko.whole}
\title{M27.betas.leuko.whole}
\format{An object of class \code{data.frame} with 27578 rows and 2 columns.}
\usage{
HM27.betas.leuko.whole
}
\description{
M27.betas.leuko.whole
}
\keyword{datasets}
