% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM27.mask}
\alias{HM27.mask}
\title{HM27.mask}
\format{An object of class \code{character} of length 2857.}
\usage{
HM27.mask
}
\description{
Vector of 2857 probe IDs recommended for masking in HM27 array based on
Zhou et al. 2017 Nucleic Acids Research.
}
\keyword{datasets}
