% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM27.ordering}
\alias{HM27.ordering}
\title{HM27 probe ordering}
\format{An object of class \code{data.frame} with 27578 rows and 6 columns.}
\usage{
HM27.ordering
}
\description{
HM27 probe chip address
}
\keyword{datasets}
