% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.controls}
\alias{HM450.controls}
\title{HM450.controls}
\format{An object of class \code{data.frame} with 850 rows and 4 columns.}
\usage{
HM450.controls
}
\description{
Data frame of 850 rows each describing chip address of a control probe,
its color channel and the control probe type in HM450.
}
\keyword{datasets}
