% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.female.clean.chrY.probes}
\alias{HM450.female.clean.chrY.probes}
\title{M450.female.clean.chrY.probes}
\format{An object of class \code{character} of length 337.}
\usage{
HM450.female.clean.chrY.probes
}
\description{
M450.female.clean.chrY.probes
}
\keyword{datasets}
