% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.female.xlinked.chrX.probes}
\alias{HM450.female.xlinked.chrX.probes}
\title{HM450.female.xlinked.chrX.probes}
\format{An object of class \code{character} of length 3797.}
\usage{
HM450.female.xlinked.chrX.probes
}
\description{
HM450.female.xlinked.chrX.probes
}
\keyword{datasets}
