% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.mapped.probes.hg19}
\alias{HM450.mapped.probes.hg19}
\title{HM450.mapped.probes.hg19}
\format{An object of class \code{GRanges} of length 467097.}
\usage{
HM450.mapped.probes.hg19
}
\description{
HM450.mapped.probes.hg19
}
\keyword{datasets}
