% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HM450.mask}
\alias{HM450.mask}
\title{HM450.mask}
\format{An object of class \code{character} of length 50186.}
\usage{
HM450.mask
}
\description{
Vector of 50186 probe IDs recommended for masking in HM450 array based on
Zhou et al. 2017 Nucleic Acids Research. MASK.general column of the
original annotation is used here.
}
\details{
Source: http://zwdzwd.github.io/InfiniumAnnotation
}
\keyword{datasets}
