% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoProbes.R
\name{sesameData_annoProbes}
\alias{sesameData_annoProbes}
\title{Annotate Probes by Probe ID}
\usage{
sesameData_annoProbes(
  Probe_IDs,
  regs = NULL,
  collapse = TRUE,
  chooseOne = FALSE,
  column = NULL,
  sep = ",",
  out_name = NULL,
  platform = NULL,
  genome = NULL,
  silent = FALSE
)
}
\arguments{
\item{Probe_IDs}{a character vector of probe IDs}

\item{regs}{a GenomicRanges::GRanges object against which
probes will be annotated, default to genes if not given}

\item{collapse}{whether to collapse multiple regs into one}

\item{chooseOne}{choose an arbitrary annotation if multiple exist}

\item{column}{which column in regs to annotate}

\item{sep}{the delimiter for collapsing}

\item{out_name}{column header of the annotation, use column if not given}

\item{platform}{EPIC, MM285 etc. will infer from Probe_IDs if not given}

\item{genome}{hg38, mm10, will infer if not given}

\item{silent}{suppress messages}
}
\value{
a GRanges with annotated column
}
\description{
Annotate Probes by Probe ID
}
\examples{
library(GenomicRanges)
regs = sesameData_getTxnGRanges("mm10")
Probe_IDs = names(sesameData_getManifestGRanges("MM285"))
anno = sesameData_annoProbes(Probe_IDs, promoters(regs), column="gene_name")
}
