% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbes.R
\name{sesameData_getAutosomeProbes}
\alias{sesameData_getAutosomeProbes}
\title{Get autosome probes}
\usage{
sesameData_getAutosomeProbes(platform = NULL, genome = NULL)
}
\arguments{
\item{platform}{'EPIC', 'HM450' etc.}

\item{genome}{hg38, mm10, ... will infer if not given.
For additional mapping, download the GRanges object from
http://zwdzwd.github.io/InfiniumAnnotation
and provide the following argument
..., genome = sesameAnno_buildManifestGRanges("downloaded_file"),...
to this function.}
}
\value{
GRanges of autosome probes
}
\description{
Get autosome probes
}
\examples{
auto_probes <- sesameData_getAutosomeProbes('Mammal40')
}
