% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{sesameData_getManifestGRanges}
\alias{sesameData_getManifestGRanges}
\title{download Infinium manifest from repositories and return GRanges}
\usage{
sesameData_getManifestGRanges(
  platform,
  genome = NULL,
  version = manifest_base_default_version,
  decoy = FALSE,
  columns = NULL
)
}
\arguments{
\item{platform}{Mammal40, MM285, EPIC, and HM450}

\item{genome}{hg38, mm10 etc.}

\item{version}{release version, default is the latest}

\item{decoy}{whether to include probes mapped to decoy sequence}

\item{columns}{additional columns to add from the manifest to mcols}
}
\value{
GRanges
}
\description{
download Infinium manifest from repositories and return GRanges
}
\examples{
gr <- sesameData_getManifestGRanges("Mammal40")
}
