% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbes.R
\name{sesameData_getProbesByChromosome}
\alias{sesameData_getProbesByChromosome}
\title{Get Probes by Chromosome}
\usage{
sesameData_getProbesByChromosome(chrms, platform = NULL, genome = NULL)
}
\arguments{
\item{chrms}{chromosomes to subset}

\item{platform}{EPIC, HM450, Mouse}

\item{genome}{hg19, hg38, or mm10, inference by default}
}
\value{
GRanges of selected probes
}
\description{
Get Probes by Chromosome
}
\examples{
Xprobes <- sesameData_getProbesByChromosome('chrX', "Mammal40")
}
