% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\name{sesameData_getTxnGRanges}
\alias{sesameData_getTxnGRanges}
\title{convert GRangesList to transcript GRanges}
\usage{
sesameData_getTxnGRanges(grl)
}
\arguments{
\item{grl}{GRangesList object}
}
\value{
a GRanges object
}
\description{
convert GRangesList to transcript GRanges
}
\examples{
full_model <- sesameDataGet("genomeInfo.mm10")$txns
txns <- sesameData_getTxnGRanges(full_model)
## get verified protein-coding
txns <- txns[(txns$transcript_type == "protein_coding" & txns$level <= 2)]

}
