% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameData_getTxnGRanges.R
\name{sesameData_getTxnGRanges}
\alias{sesameData_getTxnGRanges}
\title{convert GRangesList to transcript GRanges}
\usage{
sesameData_getTxnGRanges(genome = NULL, grl = NULL)
}
\arguments{
\item{genome}{hg38, mm10, ...}

\item{grl}{GRangesList object}
}
\value{
a GRanges object
}
\description{
convert GRangesList to transcript GRanges
}
\examples{
txns <- sesameData_getTxnGRanges("mm10")
## get verified protein-coding
txns <- txns[(txns$transcript_type == "protein_coding" & txns$level <= 2)]

}
