% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameDataCache.R
\name{sesameDataCacheAll}
\alias{sesameDataCacheAll}
\title{Cache all SeSAMe data}
\usage{
sesameDataCacheAll(data_titles = NULL)
}
\value{
TRUE
}
\description{
Cache all SeSAMe data
}
\examples{
\dontrun{
sesameDataCacheAll()
}
}
