% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sesameDataCacheAll}
\alias{sesameDataCacheAll}
\title{Cache all SeSAMe data}
\usage{
sesameDataCacheAll()
}
\value{
TRUE
}
\description{
Cache all SeSAMe data
}
\examples{
if(FALSE) { sesameDataCacheAll() }
}
