% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sesameDataCacheAll}
\alias{sesameDataCacheAll}
\title{Cache all SeSAMe data}
\usage{
sesameDataCacheAll(showProgress = FALSE)
}
\arguments{
\item{showProgress}{whether to show progress of download}
}
\value{
TRUE
}
\description{
Cache all SeSAMe data
}
\examples{
sesameDataCacheAll()
}
