% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sesameDataCacheAll}
\alias{sesameDataCacheAll}
\title{Cache all SeSAMe data}
\usage{
sesameDataCacheAll(dateAdded = latest_date)
}
\arguments{
\item{dateAdded}{version of the data by date added, if "all", cache all dates}
}
\value{
TRUE
}
\description{
Cache all SeSAMe data
}
\examples{
sesameDataCacheAll()
}
