% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameDataGet.R
\name{sesameDataGet}
\alias{sesameDataGet}
\title{Get SeSAMe data}
\usage{
sesameDataGet(title, use_alternative = FALSE, verbose = FALSE)
}
\arguments{
\item{title}{title of the data}

\item{use_alternative}{to use alternative hosting site}

\item{verbose}{whether to output ExperimentHub message}
}
\value{
data object
}
\description{
Get SeSAMe data
}
\examples{

sesameDataCacheExample()
EPIC.1.SigDF <- sesameDataGet('EPIC.1.SigDF')
}
