% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesameDataGet.R
\name{sesameDataList}
\alias{sesameDataList}
\title{List all SeSAMe data}
\usage{
sesameDataList(filter = NULL)
}
\arguments{
\item{filter}{keyword to filter title, optional}
}
\value{
all titles from SeSAMe Data
}
\description{
List all SeSAMe data
}
\examples{
sesameDataList("KYCG")
}
