% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC.refGene.gene2txn.hg38}
\alias{UCSC.refGene.gene2txn.hg38}
\title{UCSC.refGene.gene2txn.hg38}
\format{An object of class \code{list} of length 27221.}
\usage{
UCSC.refGene.gene2txn.hg38
}
\description{
List of 27221 elements each contains the transcript ID for a gene.
The data was adapted from UCSC refGene (2016) for genome build hg38.
}
\details{
See scripts/make-data.R for building details.
}
\keyword{datasets}
