% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC.refGene.txn2gene.hg38}
\alias{UCSC.refGene.txn2gene.hg38}
\title{UCSC.refGene.txn2gene.hg38}
\format{An object of class \code{list} of length 58056.}
\usage{
UCSC.refGene.txn2gene.hg38
}
\description{
UCSC.refGene.txn2gene.hg38
}
\keyword{datasets}
