% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC.refGene.txns.hg19}
\alias{UCSC.refGene.txns.hg19}
\title{CSC.refGene.txns.hg19}
\format{An object of class \code{GRangesList} of length 62083.}
\usage{
UCSC.refGene.txns.hg19
}
\description{
CSC.refGene.txns.hg19
}
\keyword{datasets}
