% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCSC.refGene.txns.hg38}
\alias{UCSC.refGene.txns.hg38}
\title{UCSC.refGene.txns.hg38}
\format{An object of class \code{GRangesList} of length 66553.}
\usage{
UCSC.refGene.txns.hg38
}
\description{
UCSC.refGene.txns.hg38
}
\keyword{datasets}
